+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - ScrollQueue - Added OnItemTrimmed event
+ - ScrollQueueTests - Added OnItemTrimmed event test
 
 ## [11.0.0] - 2020-03-20
 ### Added